pyproject.toml
[tool.poetry]
name = "rep2recall"
version = "0.4.0"
description = "Repeat until you can recall"
authors = ["Pacharapol Withayasakpunt"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-cors = "^3.0"
appdirs = "^1.4"
flask-socketio = "^4.1"
gevent = "^1.4"
python-dotenv = "^0.10.3"
pymongo = "^3.8"
dnspython = "^1.16"
python-dateutil = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
